Progress on ride request controller

+++ safe-rides-api/src/main/java/edu/csus/asi/saferides/model/RideRequest.java
@@ -76,13 +76,13 @@ public class RideRequest {
      * The odometer reading of the driver's vehicle at the beginning of the ride
      */
     @Column
-    private int startOdometer;
+    private long startOdometer;
 
     /**
      * The odometer reading of the driver's vehicle at the time of ride completion
      */
     @Column
-    private int endOdometer;
+    private long endOdometer;
 
     /**
      * Line 1 of rider's pickup address
@@ -375,7 +375,7 @@ public class RideRequest {
      *
      * @return driver's start odometer at the time of assignment
      */
-    public int getStartOdometer() {
+    public long getStartOdometer() {
         return startOdometer;
     }
 
@@ -384,7 +384,7 @@ public class RideRequest {
      *
      * @param startOdometer the driver's start odometer at the time of assignment
      */
-    public void setStartOdometer(int startOdometer) {
+    public void setStartOdometer(long startOdometer) {
         this.startOdometer = startOdometer;
     }
 
@@ -393,7 +393,7 @@ public class RideRequest {
      *
      * @return the driver's end odometer at the time of ride completion
      */
-    public int getEndOdometer() {
+    public long getEndOdometer() {
         return endOdometer;
     }
 
@@ -402,7 +402,7 @@ public class RideRequest {
      *
      * @param endOdometer the driver's end odometer at the time of ride completion
      */
-    public void setEndOdometer(int endOdometer) {
+    public void setEndOdometer(long endOdometer) {
         this.endOdometer = endOdometer;
     }
